# functions/src/openapi.yaml
openapi: 3.0.3
info:
  title: CrowdPM API
  version: 1.0.0
  description: >
    API used by CrowdPM to manage devices and retrieve pollution measurements.
  license:
    name: Proprietary
    url: https://crowdpmplatform.web.app
servers:
  - url: https://crowdpmplatform.web.app
    description: Production
  - url: http://127.0.0.1:5001/demo-crowdpm/us-central1/crowdpmApi
    description: Local development (Firebase emulator)
security: []
paths:
...
  /device/start:
    post:
      operationId: startDevicePairing
      summary: Start device pairing
      description: >
        Called by a headless node to begin the pairing flow. The node supplies its
        ephemeral Ed25519 public key (`pub_ke`) along with hardware metadata and
        receives a `device_code` + human-readable `user_code`. A human must visit
        `/activate` to approve the request before the device can continue.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pub_ke, model, version]
              properties:
                pub_ke:
                  type: string
                  description: Base64url-encoded Ed25519 public key that will be used with DPoP during pairing.
                model:
                  type: string
                  example: rpi-node
                version:
                  type: string
                  example: 1.2.3
                nonce:
                  type: string
                  description: Optional client-provided nonce for idempotency.
      responses:
        '200':
          description: Pairing session created.
          content:
            application/json:
              schema:
                type: object
                required: [device_code, user_code, verification_uri, poll_interval, expires_in]
                properties:
                  device_code:
                    type: string
                    description: 128-bit opaque identifier used by the device when polling.
                  user_code:
                    type: string
                    description: Human-readable code displayed to the operator.
                  verification_uri:
                    type: string
                    format: uri
                    description: URL where the human completes activation (`/activate`).
                  verification_uri_complete:
                    type: string
                    format: uri
                    description: Deep link that includes the user code in the query string.
                  poll_interval:
                    type: integer
                    description: Minimum seconds between token polls.
                  expires_in:
                    type: integer
                    description: Lifetime of the device and user codes in seconds.
        '400':
          description: Invalid pairing request.
        '429':
          description: Rate limited.
  /device/token:
    post:
      operationId: pollPairingToken
      summary: Poll for a registration token
      description: >
        Called by the device after the human approves pairing. Requires a DPoP proof bound
        to the `pub_ke` submitted to `/device/start`. While the human has not yet approved,
        the endpoint returns `authorization_pending`. After approval it returns a short-lived
        `registration_token` JWT that can be redeemed via `/device/register`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_code]
              properties:
                device_code:
                  type: string
      responses:
        '200':
          description: Registration token issued.
          content:
            application/json:
              schema:
                type: object
                required: [registration_token, expires_in]
                properties:
                  registration_token:
                    type: string
                    description: JWT scoped to the `device_register` audience.
                  expires_in:
                    type: integer
                    description: Token lifetime in seconds.
        '400':
          description: Session pending, expired, or invalid.
  /device/register:
    post:
      operationId: completeDeviceRegistration
      summary: Register a device identity
      description: >
        Exchanges a `registration_token` for a long-lived device identity. The caller must
        present the bearer token in `Authorization: Bearer` *and* a DPoP proof bound to the
        ephemeral key (`Ke`). The body must include the device's long-term Ed25519 key (`Kl`)
        as a JWK.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jwk_pub_kl:
                  type: object
                  description: Ed25519 JWK representing the device's long-term key.
                csr:
                  type: string
                  description: Optional PKCS#10 CSR (not yet supported).
      responses:
        '200':
          description: Device registered.
          content:
            application/json:
              schema:
                type: object
                required: [device_id, jwk_pub_kl, issued_at]
                properties:
                  device_id:
                    type: string
                  jwk_pub_kl:
                    type: object
                  issued_at:
                    type: integer
                    description: UNIX timestamp (seconds) when the device record was created.
        '400':
          description: Invalid request payload.
        '401':
          description: Missing/invalid registration token or DPoP proof.
        '403':
          description: Session not authorized for this account.
  /device/access-token:
    post:
      operationId: issueDeviceAccessToken
      summary: Issue a DPoP-bound device access token
      description: >
        Exchanges proof-of-possession of the registered long-term key (`Kl`) for a short-lived
        access token scoped to ingest APIs. Requires a DPoP proof bound to `Kl`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id]
              properties:
                device_id:
                  type: string
                scope:
                  type: array
                  items:
                    type: string
                  description: Optional scopes (defaults to `ingest.write`).
      responses:
        '200':
          description: Access token issued.
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in, device_id]
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: DPoP
                  expires_in:
                    type: integer
                  device_id:
                    type: string
        '401':
          description: Missing or invalid DPoP proof.
        '403':
          description: Device not active.
  /v1/devices:
    get:
      operationId: listDevices
      summary: List devices
      description: >
        Returns the devices the authenticated caller can access. Devices may be
        shared with multiple Firebase users, and the caller only receives entries
        where their UID is a member.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Devices returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          description: Authentication required.
    post:
      summary: Register a device
      description: >
        Registers a new device for the authenticated user. The device inherits the
        requesting user's Firebase UID as its owner.
      operationId: registerDevice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRegistrationResponse'
        '400':
          description: Invalid request payload.
        '401':
          description: Missing or invalid bearer token.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DeviceStatus:
      type: string
      enum:
        - ACTIVE
        - SUSPENDED
    Device:
      type: object
      description: A registered CrowdPM device.
      properties:
        id:
          type: string
          description: Device identifier.
        name:
          type: string
          description: Friendly name supplied by the owner.
        ownerUserId:
          type: string
          description: Primary Firebase UID associated with the device. Legacy field retained for backwards compatibility.
        ownerUserIds:
          type: array
          description: All Firebase UIDs that can access the device.
          items:
            type: string
          minItems: 1
        publicDeviceId:
          type: string
          description: Raw device identifier provided by the owner (before namespacing).
        ownerScope:
          type: string
          description: Slugged segment derived from the owner UID that is prepended to device IDs for uniqueness.
        status:
          $ref: '#/components/schemas/DeviceStatus'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the device record was created.
        location:
          $ref: '#/components/schemas/Location'
      required:
        - id
        - name
        - ownerUserId
        - status
        - createdAt
...
    Location:
      type: object
      description: Latitude and longitude in decimal degrees.
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
      required:
        - latitude
        - longitude
    RegisterDeviceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: CrowdPM Sensor - Home
          description: Friendly name to display in dashboards.
    DeviceRegistrationResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Newly generated device identifier.
...